refactor: BookLibraryJdbcRepository 에서 JdbcTemplate 대신 NamedParameterJdbcTemplate 사용하도록 변경
